More tests of UnionFindSet
Hit my first segmentation fault. The fix was to not pass root_size as a
pointer into the ufset_* functions but to instead pass it as a
reference.
I don't know why dereferncing as a pointer would cause a segfault. I'll
have to do more research.
Implementation is looking good though and using pointers and references
is feeling pretty natural so far.